chatbot-api/pyproject.toml
[project]
name = "chatbot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dos-utility",
    "fastapi[standard]>=0.128.0",
    "nh3>=0.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    # LlamaIndex core + workflow
    "llama-index-core>=0.12.0",
    "llama-index-utils-workflow>=0.3.0",
    # OpenTelemetry tracing
    "opentelemetry-sdk>=1.30.0",
    # OpenInference instrumentation for LlamaIndex
    "openinference-instrumentation-llama-index>=3.0.0",
    # LLM provider — uncomment the one you are using:
    "llama-index-llms-google-genai>=0.4.0",
    # "llama-index-llms-openai>=0.4.0",
    # "llama-index-llms-azure-openai>=0.3.0",
    # "llama-index-llms-bedrock>=0.5.0",
    # Embedding model provider — uncomment the one you are using:
    "llama-index-embeddings-google-genai>=0.3.0",
    # "llama-index-embeddings-openai>=0.3.0",
    # "llama-index-embeddings-azure-openai>=0.3.0",
    # "llama-index-embeddings-bedrock>=0.5.0",
    # Google GenAI SDK (required by models.py for EmbedContentConfig)
    "google-genai>=1.0.0",
    # uuid7 is built-in only from Python 3.14+; use backport on 3.13
    "uuid6>=2022.10.25",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pip-licenses>=5.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-dotenv>=0.5.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
env_files = [".env.test"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
dos-utility = { path = "../dos-utility" }
